Created Cell Data struct and Position classes
Created a struct called CellData that contains the value to be displayed on the cell and the color of the player. This will be used to check who gets the point when an SOS is formed in the future. The position class basically stores the x and y values of a cell and is used to assign cell data to a position on the board.