Refactoring and Computer Player Win Move
Refactored a lot of the code. Now there are two new classes. ComputerGameLogic is a class that inherits from GameLogic. It contains methods for a fully automatic SOS game, as well as methods for a computer player to make random moves and winning moves. TestingGameLogic is the other new class which inherits from ComputerGameLogic and contains all of the methods needed for testing gameLogic. The constructor for gameLogic has also been refactored so that it takes in an integer to determine the board size instead of having to use a separate method.  Multiple other methods have been refactored as well.